Improve average performance of the solver by predicting if we can finish or not.
I added two ways for us to do that.

Previously we would check if the current
solution was possible before incrementing
the current position. If the current solution
is not possible then we shouldn't bother checking
if we're done because we know we're not.

The second thing to do is to know how many blank
squares are left and check if zero are left at the
top of our loop. If currently we have a possible
solution and there are zero blanks left then we're
done.

The only difference between the isPossibleSolution
and isComplete functions is that the
isPossibleSolution does not fail if a blank is
detected. Therefor we can consider a solution found
if isPossibleSolution returns true and there are
no blank spaces left on the board.